backend/pyproject.toml
[project]
name = "doornegar-backend"
version = "0.1.0"
description = "Doornegar - Iranian Media Transparency Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "httpx>=0.27",
    "feedparser>=6.0",
    "trafilatura>=1.8",
    "langdetect>=1.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "jdatetime>=5.0",
    "pgvector>=0.3",
    "scikit-learn>=1.4",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
nlp = [
    "hazm>=0.10",
    "sentence-transformers>=3.0",
    "torch>=2.2",
    "transformers>=4.40",
]
llm = [
    "anthropic>=0.30",
    "openai>=1.30",
]
social = [
    "telethon>=1.36",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
